Corrected Logic, Retuned, and Regenerated Results
During review of the paper, we discovered that a negation was missing in the logic of CBI-2. This only affects swallowing behavior, specifically in the activation of B31/B32. In the version without negation, CBI-2 was improperly activating and thus turning on B31/B32. In the corrected logic, B31/B32 activation is now coordinated by pressure feedback for swallowing (as expected). To achieve the same model before (with only minute changes in the kinematics and timing), the pressure feedback threshold for B31/B32 was the only parameter that needed to be retuned. All other parameters are identical as are the other behaviors. This push contains the corrected and retuned model and the regenerated results.